Expose mal backend & taylor wrapper in package exports
